+++ src/componentRenderer.tsx
@@ -29,13 +29,26 @@ interface ComponentRendererConfig {
   modules: Modules
   mode: RenderMode
   serialize: (component: SerializableComponent) => void
+  Wrapper: (props) => JSX.Element
 }
 
 export function componentRenderer(config: ComponentRendererConfig) {
-  const { modules, mode = RenderMode.SERVER, serialize } = config
+  const {
+    modules,
+    mode = RenderMode.SERVER,
+    serialize,
+    Wrapper = ({ children }) => children,
+  } = config
 
   const components = Object.keys(modules).reduce((moduleMap, moduleName) => {
-    const Component = modules[moduleName]
+    const Module = modules[moduleName]
+    const getComponent = () => props => {
+      return (
+        <Wrapper>
+          <Module {...props} />
+        </Wrapper>
+      )
+    }
 
     return {
       ...moduleMap,
@@ -44,6 +57,7 @@ export function componentRenderer(config: ComponentRendererConfig) {
         if (mode === RenderMode.SERVER) {
           const mountId = props.mountId || uniqueId("stitch-component-")
           const sheet = new ServerStyleSheet()
+          const Component = getComponent()
           const html = renderToString(
             sheet.collectStyles(<Component {...props} />)
           )
@@ -56,19 +70,38 @@ export function componentRenderer(config: ComponentRendererConfig) {
           `.trim()
 
           if (serialize) {
-            serialize({ mountId, moduleName, props })
+            serialize({
+              mountId,
+              moduleName,
+              props,
+            })
           }
 
           return markup
 
           // Client
         } else if (mode === RenderMode.CLIENT) {
-          setTimeout(() => {
-            ReactDOM.hydrate(
-              <Component {...props} />,
-              document.getElementById(props.mountId)
-            )
-          }, 0)
+          setImmediate(() => {
+            if (!props.mountId) {
+              console.error(
+                "[@artsy/stitch] Error mounting clientside component: `mountId` is " +
+                  "undefined. Did you forget to pass in a `mountId` in as a prop?"
+              )
+            }
+
+            const mountPoint = document.getElementById(props.mountId)
+            if (!mountPoint) {
+              console.error(
+                "[@artsy/stitch] Error mounting client-side component: Element not " +
+                  `found for #id '${
+                    props.mountId
+                  }'. Did you forget to create a ` +
+                  "DOM node to mount react to?"
+              )
+            }
+            const Component = getComponent()
+            ReactDOM.hydrate(<Component {...props} />, mountPoint)
+          })
         }
       },
     }