+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Master
 
+- FEATURE: add ability to wrap modules with container component
 - BREAKING: Rename export of internal-use middleware export
 - Convert to TypeScript